resume/requirements.txt
transformers
torch
pandas
PyMuPDF
groq
python-dotenv
